taipy_version/requirements.txt
# ALEX Pro - Version Taipy Enterprise
taipy>=3.0.0
taipy-gui>=3.0.0
taipy-core>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-text-splitters>=0.3.9
pypdf>=3.17.0
docx2txt>=0.8
openai>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.7.0
flask>=2.0.0
pysqlite3-binary>=3.45.0